free-items-from-scheduler-py-move: Move module-level helpers out of scheduler.py
Pure mechanical relocation of module-level free items out of
``scheduler.py``. Each item is byte-identical-relocated; no rename, no
signature change.

| Item | Target |
|---|---|
| ``EmbeddingBatchResult`` | ``managers/utils.py`` (alongside ``GenerationBatchResult``) |
| ``validate_dflash_request`` | ``speculative/dflash_utils.py`` |
| ``create_scheduler_watchdog`` | ``scheduler_components/invariant_checker.py`` |
| ``IdleSleeper`` | ``scheduler_components/idle_sleeper.py`` (new file) |
| ``is_health_check_generate_req`` | ``managers/utils.py`` (fixes reverse-import from ``tokenizer_manager``) |
| ``SenderWrapper`` | ``scheduler_components/output_sender.py`` (new file) |

Caller-site impact: only import-path rewrites.

A follow-up commit (``cleanup-scheduler-py-free-items``) deletes the
dead-code ``is_work_request`` function (zero callers codebase-wide).

Refactor chain ID: free-items-from-scheduler-py-move
