+++ python/sglang/srt/managers/scheduler_components/metrics_reporter.py
@@ -32,10 +32,8 @@ from sglang.srt.utils.scheduler_status_logger import SchedulerStatusLogger
 if TYPE_CHECKING:
     from sglang.srt.managers.schedule_batch import Req
     from sglang.srt.managers.schedule_policy import PrefillAdder
-    from sglang.srt.managers.scheduler import (
-        EmbeddingBatchResult,
-        Scheduler,
-    )
+    from sglang.srt.managers.scheduler import Scheduler
+    from sglang.srt.managers.utils import EmbeddingBatchResult
 
 
 logger = logging.getLogger(__name__)
